+++ settings.py
@@ -5,6 +5,11 @@ TABLE_ID = 0
 POLLING_TIME = 3
 
 FLOW_MANAGER_URL = 'http://localhost:8181/api/kytos/flow_manager/v2'
+TOPOLOGY_URL = 'http://localhost:8181/api/kytos/topology/v3'
+
+LLDP_LOOP_ACTIONS = ["log"]  # supported actions ["log", "disable"]
+LLDP_IGNORED_LOOPS = {}  # ignored loops per dpid {"dpid": [[1, 2]]}
+LOOP_LOG_EVERY = int(max(300 / max(POLLING_TIME, 1), 1))  # 5 mins by default
 
 # Prefix this NApp has when using cookies
 COOKIE_PREFIX = 0xab